% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PxFxMean.R
\name{Fm1}
\alias{Fm1}
\alias{Fm2}
\alias{Fm3}
\alias{Fm4}
\alias{Fm1two}
\alias{Fm2two}
\alias{Fm3two}
\alias{Fm4two}
\title{Edgeworth expansion functions for sample means}
\usage{
Fm1(x, n, lam)

Fm2(x, n, lam)

Fm3(x, n, lam)

Fm4(x, n, lam)

Fm1two(x, nx, ny, lamx, lamy, varx, vary)

Fm2two(x, nx, ny, lamx, lamy, varx, vary)

Fm3two(x, nx, ny, lamx, lamy, varx, vary)

Fm4two(x, nx, ny, lamx, lamy, varx, vary)
}
\arguments{
\item{x}{a numerical value (quantile of sampling distribution) or a vector of quantiles.}

\item{n}{sample size (one-sample).}

\item{lam}{a vector of scaled cumulants or their estimates, starting with 3rd
(skewness). Term 1 needs a third cumulant only, term 2 - third and fourth
(kurtosis), and so on.}

\item{nx}{number of observations in the first group (two-sample).}

\item{ny}{number of observations in the second group (two-sample).}

\item{lamx}{scaled cumulants of the distribution of the first group (two-sample).}

\item{lamy}{scaled cumulants of the distribution of the second group (two-sample).}

\item{varx}{variance of the first group (two-sample).}

\item{vary}{variance of the second group (two-sample).}
}
\value{
A vector of values of Edgeworth expansion of a corresponding order (\code{Fm1}
  and \code{Fm1two} for a 1-term or 2nd order Edgeworth expansion and so on). The length of the vector is the same as the length of \code{x}.
}
\description{
Calculate values of 1 - 4-term Edgeworth expansions (2 - 5 order) for a sample mean and a two-sample difference in means.
}
\details{
\code{Fm1, ..., Fm4} implement Edgeworth expansions for a sample average and \code{Fm1two, ..., Fm4two} - for a two-sample difference in means. Note that for a two-sample difference in means, the first sample (\eqn{\bar{X}}) would correspond to treatment while the second (\code{Y}) -  to control.
}
\examples{

}
\seealso{
\code{\link{p1}} for \code{p()} functions used in Edgeworth expansion terms.
}
