% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QxFxTstat.R
\name{Ft1}
\alias{Ft1}
\alias{Ft2}
\alias{Ft3}
\alias{Ft4}
\alias{Ft2gen}
\alias{Ft3gen}
\alias{Ft4gen}
\title{Standalone Edgeworth expansion functions for a t-statistic}
\usage{
Ft1(x, n, lam, r = 1, norm = FALSE, df = n - 1)

Ft2(x, n, lam, r = 1, norm = FALSE, df = n - 1)

Ft3(x, n, lam, r = 1, norm = FALSE, df = n - 1)

Ft4(x, n, lam, r = 1, norm = FALSE, df = n - 1)

Ft2gen(x, n, r, k12, k22, k31, k41, norm = FALSE, df = NULL)

Ft3gen(x, n, r, k12, k13, k22, k31, k32, k41, k51, norm = FALSE, df = NULL)

Ft4gen(x, n, r, k12, k13, k22, k23, k31, k32, k41, k42, k51, k61,
  norm = FALSE, df = NULL)
}
\arguments{
\item{x}{numerical value, a quantile of sampling distribution. *** a vector
here but what about F?}

\item{lam}{a vector of scaled cumulants or their estimates, starting with 3rd
(skewness). Term 1 needs a third cumulant only, term 2 - third and fourth
(kurtosis), and so on.}

\item{r}{sqare root of variance adjustment. The variance adjustment is
generally equal to \code{k21}.}

\item{norm}{if \code{TRUE}, normal distribution is used as a base for
Edgeworth expansions, if \code{FALSE} - Student's t-distribution. Defaults
to \code{FALSE}.}

\item{df}{degrees of freedom for Student's t-distribution if \code{norm =
FALSE}. For a general version, provide a single value for the first order
approximation (zero term).}

\item{k12, k13, k22, k23, k31, k32, k41, k42, k51, k61}{cumulant components - values 
calculated from sample statistics or distribution parameters.}
}
\value{
The value of Edgeworth expansion of a corresponding order (\code{Ft1}
  and \code{Ft1gen} for a 1-term or 2nd order EE and so on). (or values, one
  for each x?)
}
\description{
Calculate values of 1 - 4-term Edgeworth expansions (2 - 5 order) for short
(ordinary one-sample) and general versions of t-statistic.
}
\details{
\code{Ft1, ..., Ft4} implement a short version of EE that can be used for 
ordinary one-sample t-statistic. Note that in this case variance adjustment
\eqn{r^2} is equal to 1 for a t-statistic with naive biased variance estimate
and \eqn{(n - 1)/n} for a standard t with unbiased variance estimate.
\code{Ft1gen, ..., Ft4gen} implement a general version, which can be used for
any t-statistic.
}
\examples{

}
\seealso{
\code{\link{q1s}}, \code{\link{q1k}} for \code{q()} functions used in
  EE terms.
}
