% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QxFxTstat.R
\name{Ftshort}
\alias{Ftshort}
\alias{Ft1}
\alias{Ft2}
\alias{Ft3}
\alias{Ft4}
\title{Standalone Edgeworth expansion for ordinary one-sample t-statistic}
\usage{
Ft1(x, n, lam, r = 1, norm = FALSE, df = n - 1)

Ft2(x, n, lam, r = 1, norm = FALSE, df = n - 1)

Ft3(x, n, lam, r = 1, norm = FALSE, df = n - 1)

Ft4(x, n, lam, r = 1, norm = FALSE, df = n - 1)
}
\arguments{
\item{x}{numeric vector of quantiles of sampling distribution.}

\item{n}{sample size.}

\item{lam}{a vector of scaled cumulants or their estimates, starting with 3rd
(skewness). Term 1 needs a third cumulant only, term 2 - third and fourth 
(kurtosis), and so on.}

\item{r}{square root of variance adjustment (see \code{Details}).}

\item{norm}{if \code{TRUE}, normal distribution is used as a base for 
Edgeworth expansions, if \code{FALSE} - Student's t-distribution. Defaults 
to \code{FALSE}.}

\item{df}{degrees of freedom for Student's t-distribution if \code{norm = 
FALSE}. A single value for the first order approximation (zero term).}
}
\value{
The value of Edgeworth expansion of a corresponding order (\code{Ft1}
  for a 1-term or 2nd order EE and so on). (or values, one for each x?)
}
\description{
Calculate values of 1 - 4-term Edgeworth expansions (EE) (2nd - 5th order) 
for ordinary one-sample t-statistic (short version).
}
\details{
Higher-order approximations of the cumulative distribution function of a 
t-statistic. These functions implement a short version of EE that can be used
for ordinary one-sample t-statistic. Note that in this case the variance 
adjustment \eqn{r^2} is equal to 1 for a t-statistic with naive biased 
variance estimate and \eqn{(n - 1)/n} for a standard t with unbiased variance
estimate.
}
\examples{

}
\seealso{
\code{\link{qfuns}} for \code{q()} functions used in EE terms and 
  \code{\link{Ftgen}} for a general version of EE. For creating EE as a 
  simple function of \code{x}, see \code{\link{makeFx}}.
}
