% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbGen.R
\name{getLam}
\alias{getLam}
\title{Scaled cumulants}
\usage{
getLam(mu)
}
\arguments{
\item{mu}{vector of 2 - 6 central moments.}
}
\value{
A named vector of 3rd to 6th scaled cumulants. The names of the
  elements are \code{"lam3", "lam4", "lam5", "lam6"}.
}
\description{
Calculate skewness, kurtosis, and 5th and 6th scaled cumulants from central
moments or their estimates.
}
