% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfuns.R
\name{Hermite}
\alias{Hermite}
\alias{He0}
\alias{He1}
\alias{He2}
\alias{He3}
\alias{He4}
\alias{He5}
\alias{He6}
\alias{He7}
\alias{He8}
\alias{He9}
\alias{He10}
\alias{He11}
\title{Probabilists' Hermite polynomials}
\usage{
He0(x)

He1(x)

He2(x)

He3(x)

He4(x)

He5(x)

He6(x)

He7(x)

He8(x)

He9(x)

He10(x)

He11(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
Probabilists' Hermite polynomials
}
