% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbmom1.R
\name{M2one}
\alias{M2one}
\title{Unbiased central moment estimates}
\usage{
M2one(m2, n)
}
\arguments{
\item{m2}{naive biased variance estimate \eqn{m_2 = 1/n \sum_{i = 1}^n ((X_i - \bar{X})^2}{m[2] = mean((X - X-bar)^2)} for a vector \code{X}.}

\item{n}{sample size.}
}
\value{
Unbiased variance estimate.
}
\description{
Calculate unbiased estimates of central moments and their powers and products.
}
\seealso{
Other unbiased estimates: \code{\link{M2M3one}},
  \code{\link{M2M3two}}, \code{\link{M2M4one}},
  \code{\link{M2M4two}}, \code{\link{M2pow2one}},
  \code{\link{M2pow2two}}, \code{\link{M2pow3one}},
  \code{\link{M2pow3two}}, \code{\link{M2two}},
  \code{\link{M3one}}, \code{\link{M3pow2one}},
  \code{\link{M3pow2two}}, \code{\link{M3two}},
  \code{\link{M4one}}, \code{\link{M4two}},
  \code{\link{M5one}}, \code{\link{M5two}},
  \code{\link{M6one}}, \code{\link{M6two}}
}
