% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbmom1.R
\name{M2pow3one}
\alias{M2pow3one}
\title{Unbiased central moment estimates}
\usage{
M2pow3one(m2, m3, m4, m6, n)
}
\arguments{
\item{m2}{naive biased variance estimate \eqn{m_2 = 1/n \sum_{i = 1}^n ((X_i 
- \bar{X})^2}{m[2] = mean((X - X-bar)^2)} for a vector \code{X}.}

\item{m3}{naive biased third central moment estimate \eqn{m_3 = 1/n \sum_{i =
1}^n ((X_i - \bar{X})^3}{m[3] = mean((X - X-bar)^3)} for a vector \code{X}.}

\item{m4}{naive biased fourth central moment estimate \eqn{m_4 = 1/n \sum_{i 
= 1}^n ((X_i - \bar{X})^4}{m[4] = mean((X - X-bar)^4)} for a vector 
\code{X}.}

\item{m6}{naive biased sixth central moment estimate \eqn{m_6 = 1/n \sum_{i =
1}^n ((X_i - \bar{X})^6}{m[6] = mean((X - X-bar)^6)} for a vector \code{X}.}

\item{n}{sample size.}
}
\value{
Unbiased estimate of cubed variance  central moment
  \eqn{\mu_2^3}{\mu[2]^3}, where \eqn{\mu_2}{\mu[2]} is a variance.
}
\description{
Calculate unbiased estimates of central moments and their powers and 
products.
}
\examples{
n <- 10
smp <- rgamma(n, shape = 3)
for (j in 2:6) {
  assign(paste("m", j, sep = ""), mean((smp - mean(smp))^j))
}
M2pow3one(m2, m3, m4, m6, n) 
}
\seealso{
Other unbiased estimates: 
\code{\link{M2M3one}()},
\code{\link{M2M3two}()},
\code{\link{M2M4one}()},
\code{\link{M2M4two}()},
\code{\link{M2one}()},
\code{\link{M2pow2one}()},
\code{\link{M2pow2two}()},
\code{\link{M2pow3two}()},
\code{\link{M2two}()},
\code{\link{M3one}()},
\code{\link{M3pow2one}()},
\code{\link{M3pow2two}()},
\code{\link{M3two}()},
\code{\link{M4one}()},
\code{\link{M4two}()},
\code{\link{M5one}()},
\code{\link{M5two}()},
\code{\link{M6one}()},
\code{\link{M6two}()}
}
\concept{unbiased estimates}
