% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbmom2.R
\name{M2two}
\alias{M2two}
\title{Pooled central moment estimates - two-sample}
\usage{
M2two(m2, n_x, n_y)
}
\arguments{
\item{m2}{naive biased variance estimate \eqn{m_2 = 1/(n_x + n_y) \sum_{i =
1}^{n_x} ((X_i - \bar{X})^2 + \sum_{i = 1}^{n_y} ((Y_i - \bar{Y})^2}{m[2] =
mean(c((X - X-bar)^2, (Y - Y-bar)^2))} for vectors \code{X} and \code{Y}.}

\item{n_x}{number of observations in the first group.}

\item{n_y}{number of observations in the second group.}
}
\value{
Pooled variance estimate.
}
\description{
Calculate pooled unbiased estimates of central moments and their powers and
products.
}
\examples{
n1 <- 10
n2 <- 8
shp <- 3
smp1 <- rgamma(n1, shape = shp) - shp
smp2 <- rgamma(n2, shape = shp)
for (j in 2:6) {
  assign(paste("m", j, sep = ""), 
         mean(c((smp1 - mean(smp1))^j, (smp2 - mean(smp2))^j)))
}          
M2two(m2, n1, n2)
}
\seealso{
Other unbiased estimates: \code{\link{M2M3one}},
  \code{\link{M2M3two}}, \code{\link{M2M4one}},
  \code{\link{M2M4two}}, \code{\link{M2one}},
  \code{\link{M2pow2one}}, \code{\link{M2pow2two}},
  \code{\link{M2pow3one}}, \code{\link{M2pow3two}},
  \code{\link{M3one}}, \code{\link{M3pow2one}},
  \code{\link{M3pow2two}}, \code{\link{M3two}},
  \code{\link{M4one}}, \code{\link{M4two}},
  \code{\link{M5one}}, \code{\link{M5two}},
  \code{\link{M6one}}, \code{\link{M6two}}
}
