% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbmom1.R
\name{M3one}
\alias{M3one}
\title{Unbiased central moment estimates}
\usage{
M3one(m3, n)
}
\arguments{
\item{m3}{naive biased third central moment estimate \eqn{m_3 = 1/n \sum_{i =
1}^n ((X_i - \bar{X})^3}{m[3] = mean((X - X-bar)^3)} for a vector \code{X}.}

\item{n}{sample size.}
}
\value{
Unbiased estimate of a third central moment.
}
\description{
Calculate unbiased estimates of central moments and their powers and 
products.
}
\examples{
n <- 10
smp <- rgamma(n, shape = 3)
for (j in 2:6) {
  assign(paste("m", j, sep = ""), mean((smp - mean(smp))^j))
}
M3one(m3, n)
}
\seealso{
Other unbiased estimates: \code{\link{M2M3one}},
  \code{\link{M2M3two}}, \code{\link{M2M4one}},
  \code{\link{M2M4two}}, \code{\link{M2one}},
  \code{\link{M2pow2one}}, \code{\link{M2pow2two}},
  \code{\link{M2pow3one}}, \code{\link{M2pow3two}},
  \code{\link{M2two}}, \code{\link{M3pow2one}},
  \code{\link{M3pow2two}}, \code{\link{M3two}},
  \code{\link{M4one}}, \code{\link{M4two}},
  \code{\link{M5one}}, \code{\link{M5two}},
  \code{\link{M6one}}, \code{\link{M6two}}
}
