% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbmom2.R
\name{M3two}
\alias{M3two}
\title{Pooled central moment estimates - two-sample}
\usage{
M3two(m3, n_x, n_y)
}
\arguments{
\item{m3}{naive biased third central moment estimate \eqn{m_3 = 1/(n_x + n_y)
\sum_{i = 1}^n_x ((X_i - \bar{X})^3 + \sum_{i = 1}^n_y ((Y_i -
\bar{Y})^3}{m[3] = mean(c((X - X-bar)^3, (Y - Y-bar)^3))} for vectors
\code{X} and \code{Y}.}

\item{n_x}{number of observations in the first group.}

\item{n_y}{number of observations in the second group.}
}
\value{
Pooled estimate of a third central moment.
}
\description{
Calculate pooled unbiased estimates of central moments and their powers and
products.
}
\seealso{
Other unbiased estimates: \code{\link{M2M3one}},
  \code{\link{M2M3two}}, \code{\link{M2M4one}},
  \code{\link{M2M4two}}, \code{\link{M2one}},
  \code{\link{M2pow2one}}, \code{\link{M2pow2two}},
  \code{\link{M2pow3one}}, \code{\link{M2pow3two}},
  \code{\link{M2two}}, \code{\link{M3one}},
  \code{\link{M3pow2one}}, \code{\link{M3pow2two}},
  \code{\link{M4one}}, \code{\link{M4two}},
  \code{\link{M5one}}, \code{\link{M5two}},
  \code{\link{M6one}}, \code{\link{M6two}}
}
