% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbmom1.R
\name{M6one}
\alias{M6one}
\title{Unbiased central moment estimates}
\usage{
M6one(m2, m3, m4, m6, n)
}
\arguments{
\item{m2}{naive biased variance estimate \eqn{m_2 = 1/n \sum_{i = 1}^n ((X_i - \bar{X})^2}{m[2] = mean((X - X-bar)^2)} for a vector \code{X}.}

\item{m3}{naive biased third central moment estimate \eqn{m_3 = 1/n \sum_{i = 1}^n ((X_i - \bar{X})^3}{m[3] = mean((X - X-bar)^3)} for a vector \code{X}.}

\item{m4}{naive biased fourth central moment estimate \eqn{m_4 = 1/n \sum_{i = 1}^n ((X_i - \bar{X})^4}{m[4] = mean((X - X-bar)^4)} for a vector \code{X}.}

\item{m6}{naive biased sixth central moment estimate \eqn{m_6 = 1/n \sum_{i = 1}^n ((X_i - \bar{X})^6}{m[6] = mean((X - X-bar)^6)} for a vector \code{X}.}

\item{n}{sample size.}
}
\value{
Unbiased estimate of a sixth central moment.
}
\description{
Calculate unbiased estimates of central moments and their powers and products.
}
\seealso{
Other unbiased estimates: \code{\link{M2M3one}},
  \code{\link{M2M3two}}, \code{\link{M2M4one}},
  \code{\link{M2M4two}}, \code{\link{M2one}},
  \code{\link{M2pow2one}}, \code{\link{M2pow2two}},
  \code{\link{M2pow3one}}, \code{\link{M2pow3two}},
  \code{\link{M2two}}, \code{\link{M3one}},
  \code{\link{M3pow2one}}, \code{\link{M3pow2two}},
  \code{\link{M3two}}, \code{\link{M4one}},
  \code{\link{M4two}}, \code{\link{M5one}},
  \code{\link{M5two}}, \code{\link{M6two}}
}
