% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQx.R
\name{makeQx}
\alias{makeQx}
\title{Make terms for Edgeworth expansions}
\usage{
makeQx(stats, type = "short", r = NULL, verbose = TRUE)
}
\arguments{
\item{stats}{named vector of distribution parameters or estimates needed for 
Edgeworth expansion. If no names for the vector are provided and the type
is \code{"short"}, elements are assumed to be scaled cumulants
\eqn{\lambda} (3 - 6). If the elements are named, the order of the elements
is arbitrary. Required names: \itemize{ \item for ordinary one-sample
t-statistic only (\code{type = "short"}): \code{"lam3", "lam4", "lam5",
"lam6"} (scaled cumulants); \item for more complicated one-sample versions,
such as a moderated t-statistic (\code{type = "one-sample"}): \code{"mu2",
"mu3", "mu4", "mu5", "mu6"} (central moments), \code{"A"}, and \code{"B"};
\item for a two-sample t-statistic (\code{type = "two-sample"}):
\code{"mu_x2", "mu_x3", "mu_x4", "mu_x5", "mu_x6"} (central moments for a
treatment group), \code{"mu_y2", "mu_y3", "mu_y4", "mu_y5", "mu_y6"}
(central moments for a control group), \code{"A", "B_x", "B_y", "b_x"}, and
\code{"b_y"}. Note that if the same distribution is assumed for the two
groups, the values for central moments for these groups should be the same
(e.g. pooled variance for a second moment). \item  optionally for a
moderated t-statistic: \code{"d0"} (prior degrees of freedom). Ignored if
\code{df} is provided. }}

\item{type}{type of Edgeworth expansions to be used: \code{"short"} for 
ordinary one-sample t-statistic, \code{"one-sample"} for a more complicated
version such as a moderated t-statistic, and \code{"two-sample"} for any 
kind of two-sample t-statistic.}

\item{r}{square root of variance adjustment. Provide if wish to use a value 
that is different from the one that will be calculated from \code{stats} 
for a general version.}

\item{verbose}{if \code{TRUE}, the warning message will be printed when the 
elements of \code{stats} are not named.}
}
\value{
A list of functions named \code{"q1"}, \code{"q2"}, \code{"q3"}, and 
  \code{"q4"} to be used in respective terms of an Edgeworth expansion 
  (orders 2 - 5). These functions take a vector of quantiles \code{x} as an 
  input and output a vector of the same length as \code{x}. For a general 
  version (\code{type = "one-sample"} and \code{type = "two-sample"}), the 
  list also includes a calculated value for variance adjustment \code{r}.
}
\description{
Create functions \code{q(x)} used in terms of Edgeworth expansions (EE) for 
various versions of t-statistic.
}
\details{
Given sample statistics or distribution parameters, produce the defining 
portion of EE terms as a function of \code{x} (quantile). Includes short (for
ordinary one-sample t-statistic) and general versions, which can be used for 
ordinary and moderated one- and two-sample t-statistics and Welch t-test. 
When used in Edgeworth expansions, the argument to a \code{q()} function 
should incorporate variance adjustment (see examples); this adjustment is
equal to 1 if naive biased variance estimates are used in a one-sample
t-statistic or a Welch test.
}
\examples{
n <- 10                 # sample size
# Gamma distribution with shape parameter \code{shp}
shp <- 3
ord <- 3:6              # orders of scaled cumulants
lambdas <- factorial(ord - 1)/shp^((ord - 2)/2)
qt <- makeQx(lambdas)
quant <- -(3:5)
qt[[1]](quant) 

# from sample
smp <- rgamma(n, shape = shp) - shp
stats <- smpStats(smp)
t <- sqrt(n)*mean(smp)/sd(smp)
r <- sqrt((n - 1)/n)
qt <- makeQx(stats)
sapply(qt, function(f) f(t/r))
# 2nd order (1-term) Edgeworth expansion, standard normal base
pnorm(t/r) + n^(-1/2)*qt[[1]](t/r)*dnorm(t/r)

}
\seealso{
\code{\link{smpStats}} that creates \code{stats} vector from the 
  sample and \code{\link{makeFx}} that creates a function returning the 
  values of Edgeworth series of orders 1 - 5 as a function of \code{x}.
}
