% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbGen.R
\name{momEdge}
\alias{momEdge}
\alias{getMomEdgeBias}
\alias{getMomEdgeUnb}
\alias{getMomEdgeBias2}
\alias{getMomEdgeUnb2}
\title{Central moment estimates for Edgeworth expansions}
\usage{
getMomEdgeBias(smp)

getMomEdgeUnb(smp)

getMomEdgeBias2(smp, a)

getMomEdgeUnb2(smp, a)
}
\arguments{
\item{smp}{sample.}

\item{a}{vector of the same length as \code{smp} specifying categories of
observations (should contain two unique values).}
}
\value{
A named vector of length \code{5}. The names of the elements
  correspond to the estimands and are \code{"mu2", "mu3", "mu4", "mu5",
  "mu6"}.
}
\description{
Calculate central moment estimates for use in Edgeworth expansions for one-
and two-sample t-tests.
}
\details{
For one-sample estimates: \code{getMomEdgeBias()} calculates regular unbiased
sample variance and naive biased estimates for 3rd to 6th central moments;
\code{getMomEdgeUnb()} calculates unbiased estimates of 2nd to 6th central
moments. For two-sample estimates, where the two populations are assumed to
have the same variance and higher central moments: \code{getMomEdgeBias2()}
calculates unbiased pooled variance and naive biased 3rd to 6th central
moments; \code{getMomEdgeUnb2()} provides unbiased pooled estimates of 2nd to
6th moments.
}
