% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PxFxMean.R
\name{p1}
\alias{p1}
\alias{p2}
\alias{p3}
\alias{p4}
\title{\code{p()} functions for Edgeworth expansions terms}
\usage{
p1(x, lam)

p2(x, lam)

p3(x, lam)

p4(x, lam)
}
\arguments{
\item{x}{a numerical value (quantile of sampling distribution) or a vector of quantiles.}

\item{lam}{a vector of scaled cumulants or their estimates, starting with 3rd
(skewness). Term 1 needs a third cumulant only, term 2 - third and fourth
(kurtosis), and so on.}
}
\value{
A vector of values of a mathematical function, one for each quantile in \code{x}.
}
\description{
Calculate values of \code{p(x)} used in terms of Edgeworth expansions (EE)
for a sample mean and a two-sample difference in means.
}
\details{
\code{p1()} is used for term 1 (second order expansion), \code{p2()} for term 2 and so on.
#'
}
\examples{

}
\seealso{
\code{\link{Fm1}}, \code{\link{Fm1two}} for corresponding EE values.
}
