% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QxFxTstat.R
\name{q1s}
\alias{q1s}
\alias{q2s}
\alias{q3s}
\alias{q4s}
\alias{q1k}
\alias{q2k}
\alias{q3k}
\alias{q4k}
\title{\code{q()} functions for Edgeworth expansions terms}
\usage{
q1s(x, lam)

q2s(x, lam)

q3s(x, lam)

q4s(x, lam)

q1k(x, r, k12, k31)

q2k(x, r, k12, k22, k31, k41)

q3k(x, r, k12, k13, k22, k31, k32, k41, k51)

q4k(x, r, k12, k13, k22, k23, k31, k32, k41, k42, k51, k61)
}
\arguments{
\item{x}{numerical value, a quantile of sampling distribution. *** a vector
here but what about F?}

\item{lam}{a vector of scaled cumulants or their estimates, starting with 3rd
(skewness). Term 1 needs a third cumulant only, term 2 - third and fourth
(kurtosis), and so on.}

\item{r}{sqare root of variance adjustment. The variance adjustment is
generally equal to \code{k21}.}

\item{k12, k13, k22, k23, k31, k32, k41, k42, k51, k61}{cumulant components - values
calculated from sample statistics or distribution parameters.}
}
\value{
The value of a mathematical function. (or values, one for each x?)
}
\description{
Calculate values of \code{q(x)} used in terms of Edgeworth expansions (EE)
for short (ordinary one-sample) and general versions of t-statistic.
}
\details{
\code{q1s, ..., q4s} implement a short version of EE that can be used for
ordinary one-sample t-statistic. \code{q1s()} is used for term 1 (second
order expansion), \code{q2s()} for term 2 and so on. They do not include
variance adjustment \code{r}, which is equal to 1 for a t-statistic with
naive biased variance estimate. For a t with unbiased variance estimate,
which is a standard, use \code{q1s(x/r)}, etc. \code{q1k, ..., q4k} implement
a general version, which can be used for any t-statistic; variance adjustment
is incorporated in these functions.
}
\examples{

}
\seealso{
\code{\link{Ft1}}, \code{\link{Ft1gen}} for corresponding EE values.
}
