% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailDiag.R
\name{tailDiag}
\alias{tailDiag}
\title{Tail diagnostic for Edgeworth expansions}
\usage{
tailDiag(stats, n, type = "short", df = NULL, moder = FALSE,
  ncheck = 50, lim = c(1, 10), verbose = TRUE)
}
\arguments{
\item{stats}{named vector of distribution parameters or estimates needed for
Edgeworth expansion. If no names for the vector are provided and the type
is \code{"short"}, elements are assumed to be scaled cumulants
\eqn{\lambda} (3 - 6). If the elements are named, the order of the elements
is arbitrary. Required names: \itemize{ \item for ordinary one-sample
t-statistic only (\code{type = "short"}): \code{"lam3", "lam4", "lam5",
"lam6"} (scaled cumulants); \item for more complicated one-sample versions,
such as a moderated t-statistic (\code{type = "one-sample"}): \code{"mu2",
"mu3", "mu4", "mu5", "mu6"} (central moments), \code{"A"}, and \code{"B"}; 
\item for a two-sample t-statistic (\code{type = "two-sample"}):
\code{"mu_x2", "mu_x3", "mu_x4", "mu_x5", "mu_x6"} (central moments for a
treatment group), \code{"mu_y2", "mu_y3", "mu_y4", "mu_y5", "mu_y6"}
(central moments for a control group), \code{"A", "B_x", "B_y", "b_x"}, and
\code{"b_y"}. Note that
if the same distribution is assumed for the two groups, the values for
central moments for these groups should be the same (e.g. pooled variance
for a second moment). \item  optionally for a moderated t-statistic: \code{"d0"}
(prior degrees of freedom). Ignored if \code{df} is provided. }}

\item{n}{a single value for a sample size summary to be used in Edgeworth
expansion. \strong{Important:} an average (not sum!) of two group sizes for a
two-sample test.}

\item{type}{which Edgeworth expansions are to be evaluated: \code{"short"}
for ordinary one-sample t-statistic, \code{"one-sample"} for a more
complicated version such as a moderated t-statistic, and
\code{"two-sample"} for any kind of two-sample t-statistic.}

\item{df}{degrees of freedom for a first order approximation, a parameter of
Student's t-distribution. If not provided, the value will be calculated
based on arguments \code{type} and \code{moder}.}

\item{moder}{\code{logical} value for calculating augmented degrees of
freedom for a moderated t-statistic; if \code{TRUE}, the value for prior
degrees of freedom shoud be included in \code{stats}. Ignored if \code{df}
is provided.}

\item{ncheck}{number of intervals for tail diagnostic.}

\item{lim}{Tail region for tail diagnostic. Provide the endpoints for the
right tail (positive values).}

\item{verbose}{if \code{TRUE}, the warning message will be printed when the
elements of \code{stats} are not named.}
}
\value{
A matrix of logical values: four columnts for orders 2 through 5, two
  rows for both tails.
}
\description{
Evaluate tail behavior and usability for 2nd - 5th order (terms 1-4)
Edgeworth expansions for a t-statistic.
}
\details{
Determine if the tails of the Edgeworth expansion behave similarly to a
cumulative distribution function and therefore can be used for approximating
sampling distribution. The results also provide information on the thickness
of the tails of that distribution (thicker tail will behave nicely). The
function is evaluated according to three criteria: monotonicity, boundedness
by 0 and 1, and being no less conservative than the first order approximation
(Student's t-distribution).

Both left and right tails are evaluated. t-distribution-based expansions are
used.
}
\examples{
# Gamma distribution with shape parameter \\code{shp}, ordinary t-statistic:
shp <- 3
n <- 10
lambda <- factorial(2:5)/shp^((1:4)/2)
tailDiag(lambda, n, "short")

# Sample statistics:
smp <- two-smp?
df.prior <- 5  # e.g. obtained from limma::eBayes()
smp.stats <- calculateStats(smp, a)
tailDiag(c(smp.stats, d0 = df.prior), mean(a), type = "two-sample", moder = TRUE)

}
\seealso{
\code{\link{smpStats}} that creates \code{stats} vector from the sample.
}
