% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PxFxMean.R
\name{Fdiff}
\alias{Fdiff}
\alias{Fm1two}
\alias{Fm2two}
\alias{Fm3two}
\alias{Fm4two}
\title{Edgeworth expansion two-sample difference in means}
\usage{
Fm1two(x, nx, ny, lamx, lamy, varx, vary)

Fm2two(x, nx, ny, lamx, lamy, varx, vary)

Fm3two(x, nx, ny, lamx, lamy, varx, vary)

Fm4two(x, nx, ny, lamx, lamy, varx, vary)
}
\arguments{
\item{x}{a numerical value (quantile of sampling distribution) or a vector of
quantiles.}

\item{nx}{number of observations in the first group (two-sample).}

\item{ny}{number of observations in the second group (two-sample).}

\item{lamx}{scaled cumulants of the distribution of the first group 
(two-sample).}

\item{lamy}{scaled cumulants of the distribution of the second group 
(two-sample).}

\item{varx}{variance of the first group (two-sample).}

\item{vary}{variance of the second group (two-sample).}
}
\value{
A vector of values of Edgeworth expansion of a corresponding order 
  (\code{Fm1} and \code{Fm1two} for a 1-term or 2nd order Edgeworth expansion
  and so on). The length of the vector is the same as the length of \code{x}.
}
\description{
Calculate values of 1 - 4-term Edgeworth expansions (EE) (2nd - 5th order)
for the two-sample standardize difference in means.
}
\details{
Higher-order approximations of the cumulative distribution function of the 
two-sample standardized difference in means. Note that for a sample \eqn{X_1,
\dots, X_{n_x}, Y_1, \dots, Y_{n_y}}{X[1], ..., X[n[x]], Y[1], ..., Y[n[y]]},
\eqn{X} would correspond to treatment and \eqn{Y} to control.
}
\seealso{
\code{\link{Fmean}} for EE for one-sample standardized mean and 
  \code{\link{pfuns}} for \code{p()} functions used in EE terms.
}
