% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PxFxMean.R
\name{Fmean}
\alias{Fmean}
\alias{Fm1}
\alias{Fm2}
\alias{Fm3}
\alias{Fm4}
\title{Edgeworth expansion for standardized mean}
\usage{
Fm1(x, n, lam)

Fm2(x, n, lam)

Fm3(x, n, lam)

Fm4(x, n, lam)
}
\arguments{
\item{x}{numeric vector of quantiles of sampling distribution.}

\item{n}{sample size.}

\item{lam}{a vector of scaled cumulants or their estimates, starting with 3rd
(skewness). Term 1 needs a third cumulant only, term 2 - third and fourth 
(kurtosis), and so on.}
}
\value{
A vector of the same length as \code{x} containing the values of 
  Edgeworth expansion of a corresponding order (\code{Fm1} for a 1-term or 
  2nd order EE, \code{Fm2} for a 2-term EE, and so on).
}
\description{
Calculate values of 1 - 4-term Edgeworth expansions (EE) (2nd - 5th order) 
for the standardized mean.
}
\details{
Higher-order approximations of the cumulative distribution function of the 
standardized sample average.
}
\examples{
# Gamma distribution with shape parameter shp
shp <- 3
ord <- 3:6                # orders of scaled cumulants
lambdas <- factorial(ord - 1)/shp^((ord - 2)/2)
x <- seq(2, 5, by = 0.5)  # thicker tail
1 - Fm1(x, n, lambdas)

}
\seealso{
\code{\link{Fdiff}} for EE for two-sample standardized difference in
  means and \code{\link{pfuns}} for \code{p()} functions used in EE terms.
}
