% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QxFxTstat.R
\name{Ftgen}
\alias{Ftgen}
\alias{Ft1gen}
\alias{Ft2gen}
\alias{Ft3gen}
\alias{Ft4gen}
\title{Standalone Edgeworth expansion for a test statistic - general case}
\usage{
Ft1gen(x, n, r, k12, k31, norm = FALSE, df = NULL)

Ft2gen(x, n, r, k12, k22, k31, k41, norm = FALSE, df = NULL)

Ft3gen(x, n, r, k12, k13, k22, k31, k32, k41, k51, norm = FALSE, df = NULL)

Ft4gen(x, n, r, k12, k13, k22, k23, k31, k32, k41, k42, k51, k61,
  norm = FALSE, df = NULL)
}
\arguments{
\item{x}{numeric vector of quantiles of sampling distribution.}

\item{k12, k13, k22, k23, k31, k32, k41, k42, k51, k61}{cumulant components - values 
calculated from sample statistics or distribution parameters.}

\item{norm}{if \code{TRUE}, normal distribution is used as a base for 
Edgeworth expansions, if \code{FALSE} - Student's t-distribution. Defaults 
to \code{FALSE}.}

\item{df}{degrees of freedom for Student's t-distribution if \code{norm = 
FALSE}. Provide a single value for the first order approximation (zero 
term).}
}
\value{
The value of Edgeworth expansion of a corresponding order 
 (\code{Ft1gen} for a 1-term or 2nd order EE and so on). (or values, one for 
 each x?)
}
\description{
Calculate values of 1 - 4-term Edgeworth expansions (EE) (2nd - 5th order) for
a general version of t-statistic and other test statistics.
}
\details{
Higher-order approximations of the cumulative distribution function of a test
statistic. These functions implement a general version of EE that can be used
for any one- or two-sample t-statistic as well as for other test statistics.
}
\examples{

}
\seealso{
\code{\link{qgen}} for \code{q()} functions used in general case EE 
 terms and \code{\link{Ftshort}} for a short version of EE. For creating EE 
 as a simple function of \code{x}, see \code{\link{makeFx}}.
}
