% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfuns.R
\name{kfuns1}
\alias{kfuns1}
\alias{K12one}
\alias{K13one}
\alias{K21one}
\alias{K22one}
\alias{K23one}
\alias{K31one}
\alias{K32one}
\alias{K41one}
\alias{K42one}
\alias{K51one}
\alias{K61one}
\title{\code{k()} functions for Edgeworth expansions - one-sample}
\usage{
K12one(A, B, mu2, mu3, mu4, mu5, mu6)

K13one(A, B, mu2, mu3, mu4, mu5, mu6)

K21one(A, B, mu2, mu3, mu4, mu5, mu6)

K22one(A, B, mu2, mu3, mu4, mu5, mu6)

K23one(A, B, mu2, mu3, mu4, mu5, mu6)

K31one(A, B, mu2, mu3, mu4, mu5, mu6)

K32one(A, B, mu2, mu3, mu4, mu5, mu6)

K41one(A, B, mu2, mu3, mu4, mu5, mu6)

K42one(A, B, mu2, mu3, mu4, mu5, mu6)

K51one(A, B, mu2, mu3, mu4, mu5, mu6)

K61one(A, B, mu2, mu3, mu4, mu5, mu6)
}
\arguments{
\item{A}{value of \code{A} (depends on the type of the test).}

\item{B}{value of \code{B} (depends on the type of the test).}

\item{mu2, mu3, mu4, mu5, mu6}{central moments (2 - 6) or their estimates.}
}
\value{
A calculated value for a respective component.
}
\description{
Calculate \code{k}'s (cumulant components) for a general version of Edgeworth
expansions (EE) for one-sample t-statistic.
}
\details{
Variance adjustment \eqn{r^2} is equal to the output of \code{K21one()},
unless different variance estimates are used for \code{A}, numerator of
\code{k}, and \code{r}.
}
\seealso{
Other \code{k()} functions: \code{\link{kfuns2}}
}
