% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QxFxTstat.R
\name{qgen}
\alias{qgen}
\alias{q1k}
\alias{q2k}
\alias{q3k}
\alias{q4k}
\title{\code{q()} functions for Edgeworth expansion terms - general version}
\usage{
q1k(x, r, k12, k31)

q2k(x, r, k12, k22, k31, k41)

q3k(x, r, k12, k13, k22, k31, k32, k41, k51)

q4k(x, r, k12, k13, k22, k23, k31, k32, k41, k42, k51, k61)
}
\arguments{
\item{x}{numeric vector of quantiles of sampling distribution.}

\item{r}{sqare root of variance adjustment. The variance adjustment is 
generally equal to \code{k21}.}

\item{k12, k13, k22, k23, k31, k32, k41, k42, k51, k61}{cumulant components - values 
calculated from sample statistics or distribution parameters.}
}
\value{
A vector of the same length as \code{x} containing the values of a 
  mathematical function.
}
\description{
Calculate values of \code{q(x)} used in terms of Edgeworth expansions (EE) 
for a general version of t-statistic and other test statistics.
}
\details{
These functions implement a general version of EE, which can be used for any
one- or two-sample t-statistic, as well as for other test statistics.
\code{q1k()} is used for term 1 (1-term expansion is also referred to as a
2nd order expansion), \code{q2k()} for term 2 and so on. Variance adjustment
is incorporated in these functions.
}
\examples{

}
\seealso{
\code{\link{Ftgen}} for corresponding EE values and 
  \code{\link{qfuns}} for the short version used in EE for ordinary
  one-sample t-statistic.
}
