% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpStats.R
\name{smpStats}
\alias{smpStats}
\title{Sample statistics for Edgeworth expansions}
\usage{
smpStats(smp, a = NULL, unbiased.mom = TRUE, type = NULL, moder = FALSE,
  d0 = NULL, s20 = NULL, varpost = NULL)
}
\arguments{
\item{smp}{sample.}

\item{a}{vector of the same length as \code{smp} specifying categories of 
observations (should contain two unique values). Treatment code is assumed
to have a higher numeric value than control (relevant for \code{type =
"Welch"}).}

\item{unbiased.mom}{\code{logical} value indicating if unbiased estimators 
for third through sixth central moments should be used.}

\item{type}{type of the test with possible values \code{"one-sample"}, 
\code{"two-sample"}, and \code{"Welch"}. For regular one- and two-sample 
tests the value is inferred from \code{a} but for Welch t-test it needs to 
be specified.}

\item{moder}{\code{logical} value indicating if Edgeworth expansions for a
moderated t-statistic will be used. If \code{TRUE}, prior information
(\code{d0} and \code{s20}) and posterior variance should be provided.}

\item{d0}{prior degrees of freedom (needed if \code{moder = TRUE}).}

\item{s20}{prior value for variance (needed if \code{moder = TRUE}).}

\item{varpost}{posterior variance (needed if \code{moder = TRUE}).}
}
\value{
A named vector of sample statistics to be used in Edgeworth
  exansions. The calculated statistics and corresponding names are: \itemize{
  \item for ordinary one-sample t-statistic: scaled cumulants named
  \code{"lam3", "lam4", "lam5", "lam6"}; \item for moderated one-sample
  t-statistic: central moment estimates named \code{"mu2", "mu3", "mu4",
  "mu5", "mu6"}, \code{A}, \code{B}, and prior degrees of freedom named
  \code{"d0"}; \item for ordinary two-sample t-statistic: central moment
  estimates repeated twice since the same distribution is assumed for two
  groups, named \code{"mu_x2", "mu_x3", "mu_x4", "mu_x5", "mu_x6"} and
  \code{"mu_y2", "mu_y3", "mu_y4", "mu_y5", "mu_y6"}, \code{A, B_x, B_y,
  b_x}, and \code{b_y}; \item  for moderated two-sample t-statistic:
  estimates of the same quantities as for ordinary t (with different
  estimators); additionally, prior degrees of freedom named \code{"d0"} is
  included; \item for Welch t-test: estimates of the same quantities as for
  ordinary t-statistic (with different estimators). In this case, central
  moment estimates for treatment and control groups are different.}
}
\description{
Calculate sample statistics needed for Edgeworth expansions.
}
\seealso{
\code{\link{tailDiag}}, \code{\link{makeFx}}, and
  \code{\link{makeQx}} for functions that use \code{stats} argument
  corresponding to the output of \code{smpStats()}.
}
